Separate the 'inTeam' option from real team status
Now the actual team status of the applicant is returned separately to the team status that the hacker indicated on their application form i.e. we can now see the separate pieces of information: did the user say they were in a team(appliedAsTeam)? did the user successfully make a team application (inTeam)?